requirements.txt
jupyter
requests
pandas
numpy
openpyxl
opencv-python
pillow
matplotlib
scipy
pytesseract
torch
torchvision
ultralytics
label-studio-sdk
python-dotenv
scikit-learn
beautifulsoup4
selenium
pdf2image
openai